requirements.txt
psycopg2-binary>=2.9.0
boto3>=1.26.0
pandas>=1.5.0,<2.1.0
scikit-learn>=1.0.0,<1.4.0
numpy>=1.21.0,<1.27.0
python-dotenv>=1.0.0
beautifulsoup4>=4.11.0
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
